[symex] Move goto_symext into goto-symex/engine/
Last of seven commits partitioning src/goto-symex/
(docs/roadmap/goto-symex-partition-plan.md §2). Everything that defines a
goto_symext member — the driver, the statement handlers, dynamic_allocation and
the builtin_functions/ sub-partition — moves under engine/. src/goto-symex/
now holds only CMakeLists.txt, .gitignore, symex_invariant.{h,cpp} and the
seven group directories.

Repoints the prose citations §9 names and re-derives LTL.md's symex_main line,
which was stale before this PR. Also records the partition in CLAUDE.md's
Source Architecture listing, the one place an agent looks for the layout.
