+++ src/esbmc/parseoptions/process_goto_program.cpp
@@ -2,7 +2,7 @@
 
 #include <esbmc/bmc.h>
 #include <esbmc/esbmc_parseoptions.h>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <goto-symex/trace/goto_trace.h>
 #include <goto-symex/trace/sarif.h>
 #include <util/base/cwe_mapping.h>
