+++ src/goto-symex/engine/builtin_functions/io.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <climits>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <goto-symex/trace/printf_formatter.h>
 #include <string>
 #include <util/arith/arith_tools.h>
