+++ src/goto-symex/engine/builtin_functions/memory_alloc.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <goto-programs/destructor.h>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <string>
 #include <util/arith/arith_tools.h>
 #include <util/lang/c_types.h>
