+++ src/goto-symex/engine/builtin_functions/threads.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <goto-symex/scheduler/execution_state.h>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <goto-symex/scheduler/reachability_tree.h>
 #include <string>
 #include <util/arith/arith_tools.h>
