+++ src/goto-symex/engine/builtin_functions/va_arg.cpp
@@ -1,4 +1,4 @@
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <string>
 #include <util/arith/arith_tools.h>
 #include <util/lang/c_types.h>
