+++ src/goto-symex/engine/builtin_functions/witness.cpp
@@ -1,4 +1,4 @@
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <goto-symex/witness/witnesses.h>
 #include <util/arith/arith_tools.h>
 #include <util/expr/expr_util.h>
