+++ src/goto-symex/engine/dynamic_allocation.h
