+++ src/goto-symex/engine/goto_symex.h
