+++ src/goto-symex/engine/symex_other.cpp
@@ -1,5 +1,5 @@
 #include <cassert>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <util/expr/expr_util.h>
 #include <irep2/irep2.h>
 #include <util/symtab/pretty.h>
