+++ src/goto-symex/engine/symex_valid_object.cpp
@@ -1,5 +1,5 @@
-#include <goto-symex/dynamic_allocation.h>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/dynamic_allocation.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <util/irep/std_expr.h>
 
 static const expr2tc *get_object(const expr2tc &expr)
