+++ src/goto-symex/equation/symex_target_equation.cpp
@@ -1,7 +1,7 @@
 #include <unordered_set>
 #include <cassert>
 #include <functional>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <goto-symex/state/goto_symex_state.h>
 #include <goto-symex/symex_invariant.h>
 #include <goto-symex/equation/symex_target_equation.h>
