+++ src/goto-symex/scheduler/reachability_tree.cpp
@@ -1,5 +1,5 @@
 #include <goto-programs/goto_functions.h>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <goto-symex/scheduler/reachability_tree.h>
 #include <irep2/irep2_expr.h>
 #include <util/config/config.h>
