+++ src/goto-symex/state/goto_symex_state.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <goto-symex/scheduler/execution_state.h>
-#include <goto-symex/goto_symex.h>
+#include <goto-symex/engine/goto_symex.h>
 #include <goto-symex/state/goto_symex_state.h>
 #include <goto-symex/scheduler/reachability_tree.h>
 #include <map>
